Example reduction in number of cogs.
