+++ src/main.py
@@ -65,8 +65,7 @@ async def on_connect():
     # bot.load_extension("commands.link_ps2_discord")
     bot.load_extension("commands.squad_markup")
     bot.load_extension("commands.ps2_lookup")
-    bot.load_extension("commands.arma_upload_map")
-    bot.load_extension("commands.arma_get_db")
+    bot.load_extension("commands.arma_cog")
     # bot.load_extension("loggers.discord_logger")
     # bot.load_extension("loggers.ps2_outfit_members")
     # bot.load_extension("loggers.ps2_outfit_online_logger")
